socket ... accept -> done

+++ core_srv/include/Server.hpp
@@ -2,14 +2,15 @@
 
 #include <iostream>
 #include <sys/socket.h>
+#include <netinet/in.h>
 
 
 class Server
 {
     private:
-        int connection; 
+        int connection;
     public:
         //          for socket fun               //        for sockaddr_in
-        Server(int domain, int type, int protocol, int port , int ip_add);
+        Server(int domain, int type, int protocol, int port , u_long ip_add);
 
 };
